==> Scripts/1_importing_libraries_and_python_data_types.ipynb <==
# 01. Importing libraries 

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Data types

# Three different ways of reching a result of 100
x = 45
y = 55
z = 10

# First way
x + y

# Second way
(y - x) * 10

# Third way
y * 2 - z

# Floating point variables
x = 1.5
y = 2.5

# Dividing them
x / y 

# Separate strings
a = "He"
b = " likes"
c = " basketball"
d = "She"
e = " swimming"

# First short sentences
a + b + c

# Second short sentences
d + b + e
