==> Scripts/2_data_import.ipynb <==
# 01. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing data 

# Import data
path = r"C:\Users\enthu\Desktop\Bootcamps\Instacart Basket Analysis"

path

df = pd.read_csv(os.path.join(path, "Data", "Original Data", "orders.csv"), index_col = False)

# A list of all columns
df.columns

# A list of the included columns
vars_list = ["order_id", "user_id", "order_number", "order_dow", "order_hour_of_day", "days_since_prior_order"]

vars_list

# Import data with only the needed columns
df = pd.read_csv(os.path.join(path, "Data", "Original Data", "orders.csv"), usecols = vars_list)

# First 5 rows
df.head(5)

# Import the second set of data
df_prods = pd.read_csv(os.path.join(path, "Data", "Original Data", "products.csv"), index_col = False)

# First 20 rows
df_prods.head(20)

#Last 35 rows
df_prods.tail(35)

# A list of all columns
df_prods.columns

# The number of rows and columns
df_prods.shape

# 03. What’s the maximum value of the “aisle_id” column?

# Descriptive statistics
df_prods.describe()

Answer: the maximum value is 134.

# 04. What’s the data type of the “department_id” column?

# Data types
df_prods.dtypes

Answer: the data type is int64.
