==> Scripts/3_data_wrangling.ipynb <==
# 01. Importing libraries

# importing libraries
import pandas as pd
import numpy as np
import os

# 02. Importing data

# importing data
path = r"C:\Users\enthu\Desktop\Bootcamps\Instacart Basket Analysis"

path

# importing orders.csv
df_ords = pd.read_csv(os.path.join(path, "Data", "Original Data", "orders.csv"), index_col = False)

# importing products.csv
df_prods = pd.read_csv(os.path.join(path, "Data", "Original Data", "products.csv"), index_col = False)

# 03. Data wrangling

# dropping eval_set column from orders.csv
df_ords.drop(columns = ["eval_set"])

# dropping the eval_set column from the actual data
df_ords_2 = df_ords.drop(columns = ["eval_set"])

# performing a value count
df_ords["days_since_prior_order"].value_counts(dropna = False)

# changing a column name
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

# changing the order_id data type to string
df_ords['order_id'] = df_ords['order_id'].astype('str')

# checking the data type of order_id
df_ords["order_id"].dtype

# importing departments.csv
df_dep = pd.read_csv(os.path.join(path, "Data", "Original Data", "departments.csv"), index_col = False)

df_dep.head()

# transposing data
df_dep.T

df_dep_t = df_dep.T

df_dep_t

# adding an index column
df_dep_t.reset_index()

# creating an object that contains the first row of the dataframe
new_header = df_dep_t.iloc[0]

new_header

# copying everything from df_dep_t to df_dep_t_new except for row 0
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# setting the header row as the df header
df_dep_t_new.columns = new_header

df_dep_t_new

# turning df_dep_t_new into a data dictionary
data_dict = df_dep_t_new.to_dict("index")

data_dict

df_prods.head()

# determining what 19 stands for
print(data_dict.get("19"))

# indexing department_id = 19
df_prods["department_id"]==19

# creating a list of department_id = 19
df_prods[df_prods["department_id"]==19]

# creating a new dataframe that cointains department_id = 19
df_snacks =  df_prods[df_prods['department_id']==19]

df_snacks.head()

# the second way of creating a new dataframe that cointains department_id = 19
df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks.head()

# the third way of creating a new dataframe that cointains department_id = 19
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

df_snacks.head()

# 0.4 Exercise 4.4

# changing the data type of user_id
df_ords['order_number'] = df_ords['order_number'].astype('str')

# double checking the data type of user_id
df_ords["order_number"].dtype

# renaming the days_since_prior_order column without overwriting the dataframe
df.rename(columns = {'days_since_prior_order' : 'days_since_last_order'}, inplace = False)

# descriptive analytics of the dataframe
df_ords.describe()

# finding the frenquency of the variables
df_ords['order_hour_of_day'].value_counts(dropna = False)

10am is the busiest hour.

# the data dictionary of the df_prods dataframe
data_dict

# finding out the meaning of department_id = 4
print(data_dict.get('4'))

# creating a list of department_id = 14
df_prods[df_prods['department_id']== 14 ]

# creating a list of department_id = 5,7,12,20
df_prods[df_prods['department_id'].isin([5,7,12,20])]

It has 7650 rows.

# creating a list of user_id = 1
df_ords[df_ords['user_id']== 1 ]

# exporting the df_ords dataframe
df_ords.to_csv(os.path.join(path, 'Data','Prepared Data', 'orders_wrangled.csv'))

# exporting the df_dep_t_new dataframe
df_dep_t_new.to_csv(os.path.join(path, 'Data','Prepared Data', 'departments_wrangled.csv'))
