==> Scripts/4_data_consistency_check.ipynb <==
# 01. Importing Libraries

# importing libraries
import pandas as pd
import numpy as np
import os

# 02. Importing Data

# importing data
path = r"C:\Users\enthu\Desktop\Bootcamps\Instacart Basket Analysis"

path

# importing products.csv
df_prods = pd.read_csv(os.path.join(path, "Data", "Original Data", "products.csv"), index_col = False)

# importing orders_wrangled.csv
df_ords = pd.read_csv(os.path.join(path, "Data", "Prepared Data", "orders_wrangled.csv"), index_col = False)

# 03. Consistency Checks

# creating a dataframe for testing
df_test = pd.DataFrame()

# crteating a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# checking for mixed data types
for col in df_test.columns.tolist():
  weird = (df_test[[col]].map(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# checking for missing values
df_prods.isnull().sum()

# creating a new dataframe that contains null values
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# number of rows in df_prods dataframe
df_prods.shape

# creating a new dataframe without missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# checking the number of columns and rows that the df_prods_clean dataframe has
df_prods_clean.shape

# creating a new dataframe that contains only duplicate rows
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

# checking the current number of rows
df_prods_clean.shape

# creating a new dataframe that doesn't include the duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# 04. Exercise 4.5

# descripotive analytics
df_ords.describe()

1. All columns seem to have the same count except for days_since_prior_order.
2. The min and max of orders_day_of_week are 0 and 6. They could represent Monday and Sunday, which is not too intuitive.
3. The min of order_hour_of_day is o, which doesn't make sense.

# looking for mixed data types
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

There are no mixed data types.

# looking for missing values in df_ords
df_ords.isnull().sum()

days_since_prior_order has 206209 missing values. Could have been caused by customers' first orders. This issue should be addressed since it's such a huge amount of missing data.

# creating a new dataframe that contains only NaN values
df_nan = df_ords[df_ords["days_since_prior_order"].isnull() == True]

df_nan

# creating a new dataframe that only contains order_number = 1
df_order_number_1 = df_ords[df_ords["order_number"] == 1]

df_order_number_1

# checking the miss values that the new dtaframe has
df_order_number_1.isnull().sum()

Both df_ords and df_order_number_1 have the same amount of missing values, which shows that the missing values were caused by first time customers. I would add a column that acts as a flag that addresses first orders.

# adding a new column to the dataframe that identifies first orders
df_ords['first_order'] = df_ords['days_since_prior_order'].isna()

df_ords

# looking for duplicates
df_ords_dups = df_prods[df_prods.duplicated()]

df_ords_dups

df_ords doesn't have any duplicate values that need to be addressed.

# exporting df_prods
df_prods.to_csv(os.path.join(path, 'Data','Prepared Data', 'prods_consistensy_checked.csv'))

# exporting df_ords
df_ords.to_csv(os.path.join(path, 'Data','Prepared Data', 'orders_consistency_checked.csv'))
