==> Scripts/5_combining_and_exporting_data.ipynb <==
## 1. Importing Data

# Import libraries
import pandas as pd
import numpy as np
import os

# Define the path
path = r"C:\Users\enthu\Desktop\Bootcamps\Instacart Basket Analysis"

path

# Import orders_products_combined pickle file
df_ords_prods_combined = pd.read_pickle(os.path.join(path, 'Data', 'Prepared Data', 'orders_products_combined.pkl'))

# Check the shape of the dataframe
df_ords_prods_combined.shape

df_ords_prods_combined.head()

# Import the products data set 
df_prods = pd.read_csv(os.path.join(path, "Data", "Prepared Data", "prods_consistensy_checked.csv"), index_col = False)

df_prods.head()

## 2. Merging Data

# Testing of merging df_ords_prods_combined dataframe and df_prods dataframe
df_merged_testing = df_ords_prods_combined.merge(df_prods, on = 'product_id')

df_merged_testing.head()

"Unnamed: 0_x" and "Unnamed: 0_y" columns need to be dropped. 

#Drop unnamed： 0 column in df_prods
df_prods = df_prods.drop(columns = ['Unnamed: 0'])

#Drop unnamed： 0 column in df_ords_prods_combined
df_ords_prods_combined = df_ords_prods_combined.drop(columns = ['Unnamed: 0'])

# Final merge after dropping columns
df_merged = df_ords_prods_combined.merge(df_prods, on = 'product_id')

df_merged.head()

df_merged['_merge'].value_counts()

# Export data to pkl
df_merged.to_pickle(os.path.join(path, 'Data','Prepared Data', 'ords_prods_merge.pkl'))
