==> Scripts/6_combining_data.ipynb <==
# 4.6 Merging and Exporting Data

## 1. Importing Data

# Import libraries
import pandas as pd
import numpy as np
import os

## 2. Concatenating Data

# Define a dictionary containing January 2020 data
data1 = {'customer_id':['6732', '767', '890', '635'],
    'month':['Jan-20', 'Jan-20', 'Jan-20', 'Jan-20'],
    'purchased_meat':[0, 13, 3, 4],
    'purchased_alcohol':[1, 2, 10, 0],
    'purchased_snacks':[10, 5, 1, 7]}

# Define a dictionary containing February 2020 data
data2 = {'customer_id':['6732', '767', '890', '635'],
    'month':['Feb-20', 'Feb-20', 'Feb-20', 'Feb-20'],
    'purchased_meat':[0, 10, 5, 3],
    'purchased_alcohol':[2, 4, 14, 0],
    'purchased_snacks':[15, 3, 2, 6]}

#Convert the dictionaries into dataframes
df = pd.DataFrame(data1,index=[0, 1, 2, 3])
df_1 = pd.DataFrame(data2,index=[0, 1, 2, 3])

df

df_1

# Create a list
frames = [df, df_1]

# Concatenate the dataframes
df_concat = pd.concat(frames)

df_concat

# Concatenate the dataframes side by side
df_concat = pd.concat(frames, axis = 1)

df_concat

## 3. Merging Data

# Define another dictionary
data3 = {'customer_id':['6732', '767', '890', '635'],
    'month':['Jan-20', 'Jan-20', 'Jan-20', 'Jan-20'],
    'days_purchased_on':[0, 10, 4, 1]}

# Create another dataframe
df_2 = pd.DataFrame(data3,index=[0, 1, 2, 3])

# Merge the dataframes using customer_id
df_merged = df.merge(df_2, on = 'customer_id')

df_merged

# Merge the dataframes using customer_id and month
df_merged = df.merge(df_2, on = ['customer_id', 'month'])

df_merged

# Merge the dataframes using customer_id and monthe with the indicator
df_merged = df.merge(df_2, on = ['customer_id', 'month'], indicator = True)

df_merged

df_merged['_merge'].value_counts()

# Test merge without overwriting
pd.merge(df,df_2, on = ['customer_id', 'month'], indicator = True)

# Define the path
path = r"C:\Users\enthu\Desktop\Bootcamps\Instacart Basket Analysis"

path

# Import the orders_products_prior file
df_ords_prior = pd.read_csv(os.path.join(path, 'Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

df_ords_prior.head()

# Import the orders_wrangled file
df_ords = pd.read_csv(os.path.join(path, "Data", "Prepared Data", "orders_wrangled.csv"), index_col = False)

df_ords.head()

# Check the shape of df_ords_prior
df_ords_prior.shape

# Check the shape of df_ords
df_ords.shape

# Merge the two dataframes using order_id
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# Check the columns of the merge
df_merged_large.head()

df_merged_large['_merge'].value_counts()

# Export data to pkl
df_merged_large.to_pickle(os.path.join(path, 'Data','Prepared Data', 'orders_products_combined.pkl'))
