==> Scripts/7_deriving_new_variables.ipynb <==
# 1. Creating the "price_label" and "busiest_day" Columns

# Import libraries
import pandas as pd
import numpy as np
import os

# Define the path to import data
path = r"C:\Users\enthu\Desktop\Bootcamps\Instacart Basket Analysis"

path

# Import ords_prods_merge
df_ords_prods_merge = pd.read_pickle(os.path.join(path, 'Data', 'Prepared Data', 'ords_prods_merge.pkl'))

# Create a subset of the dataframe
df = df_ords_prods_merge[:1000000]

df

# Set up a function called price_label
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Apply the function
df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

# Check the most expensive product
df['prices'].max()

# Set up a loc function
df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'
df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 
df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

# Run a loc function on the entire dataframe
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'
df_ords_prods_merge.loc[(df_ords_prods_merge['prices'] <= 15) & (df_ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_merge['price_range_loc'].value_counts(dropna = False)

# Find out which day most orders take place
df_ords_prods_merge['orders_day_of_week'].value_counts(dropna = False)

Most orders take place on a Saturday.

# Create an empty list that can contain different results
result = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

# Create a new column with the new information
df_ords_prods_merge['busiest_day'] = result

df_ords_prods_merge['busiest_day'].value_counts(dropna = False)

# 2. Exercise 4.7

# 0 and 1 are the busiest days. 3 and 4 are the slowest days. Create another empty list that can contain the new required results
result_2 = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0 or value == 1:
    result_2.append("Busiest days")
  elif value == 3 or value == 4:
    result_2.append("Slowest days")
  else:
    result_2.append("Regularly busy")

result_2

# Create a new column with the new information
df_ords_prods_merge['busiest_days'] = result_2

df_ords_prods_merge['busiest_days'].value_counts(dropna = False)

df_ords_prods_merge.head()

df_ords_prods_merge['busiest_day'].count()

df_ords_prods_merge['busiest_days'].count()

Both the busiest_day column and the busiest_days column have the same count, which means the for loops were run correctly. Busiest days and regularly busy days are a huge portion of the total days. It is a good sign beacuase the business is running well. 

# Find out wchich hour most orders take place
df_ords_prods_merge['order_hour_of_day'].value_counts(dropna = False)

The hours should be divided into three periods equally, which are "Most orders", "Average orders", and "Fewest orders".

# Create a for loop for different periods
result_3 = []

for value in df_ords_prods_merge["order_hour_of_day"]:
  if value in [10, 11, 14, 15, 13, 12, 16, 9]:
    result_3.append("Most orders")
  elif value in [17, 8, 18, 19, 20, 7, 21, 22]:
    result_3.append("Average orders")
  else:
    result_3.append("Fewest orders")

result_3

# Create a new column with the new information
df_ords_prods_merge['busiest_period_of_day'] = result_3

df_ords_prods_merge['busiest_period_of_day'].value_counts(dropna = False)

# Export the dataframe as a pickle file
df_ords_prods_merge.to_pickle(os.path.join(path, 'Data', 'Prepared Data', 'orders_products_merged_derived.pkl'))
