==> Scripts/9_grouping_data_&_aggregating_variables_part_2.ipynb <==
## 1. Importing Data

# Import libraries
import pandas as pd
import numpy as np
import os

# Define the path to import data
path = r"C:\Users\enthu\Desktop\Bootcamps\Instacart Basket Analysis"

path

# Import ords_prods_merge
ords_prods_merge = pd.read_pickle(os.path.join(path, 'Data', 'Prepared Data', 'orders_products_merged_derived.pkl'))

## 2. Grouping and Aggregating Data

# Get the means of the "order_number" column and split them into groups based on "department_id"
ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

Since the subset was only the first one million rows, it was missing a lot of depatment_id. The entire dataframe also has different means because it includes much more rows.

# Create a column named max_order used to store the maximum order_number grouped bny user_id
ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

ords_prods_merge.head(15)

# Create a new column named loyalty_flag and assign different flags to it
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

# Get the means, mins, and maxes of the order_number column and split them into groups based on "loyalty_flag"
ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max']})

The prices of products purchased by different customers differ from each other.

# Create a column named average_price used to store the averages of prices grouped by user_id
ords_prods_merge['average_price'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

ords_prods_merge.head(15)

# Create a new column named spenders and assign different flags to it
ords_prods_merge.loc[ords_prods_merge['average_price'] < 10, 'spenders'] = 'Low spender'
ords_prods_merge.loc[ords_prods_merge['average_price'] >= 10, 'spenders'] = 'High spender'

# Check the amount of different spenders
ords_prods_merge['spenders'].value_counts(dropna = False)

# Create a column named frequency_median used to store the medians of frequency grouped bny user_id
ords_prods_merge['frequency_median'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

ords_prods_merge.head(15)

# Create a new column named customers and assign different flags to it
ords_prods_merge.loc[ords_prods_merge['frequency_median'] > 20, 'customers'] = 'Non-frequent customer'
ords_prods_merge.loc[(ords_prods_merge['frequency_median'] > 10) & (ords_prods_merge['frequency_median'] <= 20), 'customers'] = 'Regular customer'
ords_prods_merge.loc[ords_prods_merge['frequency_median'] <= 10, 'customers'] = 'Frequent customer'

# Check the amount of different types of customers
ords_prods_merge['customers'].value_counts(dropna = False)

ords_prods_merge.to_pickle(os.path.join(path, 'Data', 'Prepared Data', 'orders_products_grouped_aggregated.pkl'))
