==> Scripts/11_python_data_visualization_part_1.ipynb <==
## 1. Importing Data

# Import libraries                  
import pandas as pd
import numpy as np
import os

# Define the path to import data
path = r"C:\Users\enthu\Desktop\Bootcamps\Instacart Basket Analysis"

path

# Import the customers data set
customers = pd.read_csv(os.path.join(path, 'Data', 'Original Data', 'customers.csv'), index_col = False)

## 2. Consistency Checks

customers.head()

There are some spelling mistakes and a few column names are not too intuitive that need to be changed. Not dropping columns because they all seem to be useful.

# Change First Name to first_name
customers.rename(columns = {'First Name' : 'first_name'}, inplace = True)

# Change Surnam to surname
customers.rename(columns = {'Surnam' : 'surname'}, inplace = True)

# Change Gender to gender
customers.rename(columns = {'Gender' : 'gender'}, inplace = True)

# Change STATE to state
customers.rename(columns = {'STATE' : 'state'}, inplace = True)

# Change Age to age
customers.rename(columns = {'Age' : 'age'}, inplace = True)

# Double check the column names
customers.head()

# descriptive analytics
customers.describe()

# Check the shape of the dataframe
customers.shape

# Check the data types
customers.dtypes

# Check whether the data set has mixed data types
for col in customers.columns.tolist():
  weird = (customers[[col]].map(type) != customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (customers[weird]) > 0:
    print (col)

# Convert the data type of the first_name column to string
customers['first_name'] = customers['first_name'].astype('str')

# Look for mixed data types again
for col in customers.columns.tolist():
  weird = (customers[[col]].map(type) != customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (customers[weird]) > 0:
    print (col)

There are no more mixed data types.

# Look for missing values
customers.isnull().sum()

There are no missing values.

# Look for dubplicates
customers_dups = customers[customers.duplicated()]

customers_dups

There are no duplicates.

## 3. Merging Data

# Import the latest prepared data file
ords_prods = pd.read_pickle(os.path.join(path, 'Data', 'Prepared Data', 'orders_products_grouped_aggregated.pkl'))

ords_prods.head()

customers.dtypes

ords_prods.dtypes

Key columns have the same data type.

# Merge ords_prods with customers using user_id
cstms_ords_prods = ords_prods.merge(customers, on = 'user_id')

cstms_ords_prods.head()

# Check the shape of the new merged dataframe
cstms_ords_prods.shape

# Check the data types of the new dataframe
cstms_ords_prods.dtypes

# Export the dataframe as a pickle file
cstms_ords_prods.to_pickle(os.path.join(path, 'Data','Prepared Data', 'customers_orders_products_merged.pkl'))
